move attr filtering macro
